pyproject.toml
[project]
name = "p12-sport-data-solution"
version = "0.1.0"
description = "POC pipeline données sportives - Sport Data Solution"
authors = [
    {name = "Alexandre Quintin",email = "alexandre.quintin@chill-data.com"}
]
readme = "README.md"
requires-python = "^3.11.8"
dependencies = ["pandas (>=3.0.1,<4.0.0)", "openpyxl (>=3.1.5,<4.0.0)", "numpy (>=2.4.3,<3.0.0)", "psycopg2-binary (>=2.9.11,<3.0.0)", "pydantic (>=2.12.5,<3.0.0)", "python-dotenv (>=1.2.2,<2.0.0)", "sqlalchemy (>=2.0.48,<3.0.0)", "requests (>=2.32.5,<3.0.0)", "confluent-kafka (>=2.13.2,<3.0.0)"]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
ipykernel = "^7.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
